part3/requirements.txt
beautifulsoup4
aiohttp
asyncio
google-cloud-pubsub
pandas
lxml